requirements.txt
PyQt5>=5.15.0
httpcore>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
pyyaml>=6.0
loguru>=0.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0

# 测试依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
